requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
wordcloud>=1.9.0
jieba>=0.42.1
tkinter>=8.6